Simplify config
