+++ src/main/groovy/io/seqera/wave/auth/RegistryLookupServiceImpl.groovy
@@ -27,7 +27,6 @@ import com.google.common.cache.CacheLoader
 import com.google.common.cache.LoadingCache
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
-import io.seqera.wave.auth.cache.RegistryAuthCacheStore
 import io.seqera.wave.configuration.HttpClientConfig
 import io.seqera.wave.http.HttpClientFactory
 import io.seqera.wave.util.Retryable
