+++ src/test/groovy/io/seqera/wave/auth/RegistryAuthServiceTest.groovy
@@ -26,7 +26,6 @@ import spock.lang.Specification
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.annotation.Value
 import io.micronaut.test.extensions.spock.annotation.MicronautTest
-import io.seqera.wave.auth.cache.RegistryTokenCacheStore
 import io.seqera.wave.test.SecureDockerRegistryContainer
 import jakarta.inject.Inject
